pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brinks-box-trading"
version = "0.1.0"
description = "Brinks Box Hybrid System — Automated Trading Engine"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "alpaca-py>=0.21",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "websockets>=12.0",
    "httpx>=0.27",
    "aiohttp>=3.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
